Dev mode refactoring
