+++ modules/core/json-object.ts
@@ -1,0 +1,7 @@
+export type JsonObject =
+    | string
+    | number
+    | boolean
+    | null
+    | { [property: string]: JsonObject }
+    | JsonObject[];
