+++ modules/core/runner/bridge-constructor.resolver.ts
@@ -1,10 +1,11 @@
 import { Constructor } from "@core/constructor";
+import { JsonObject } from "@core/json-object";
 import { IRunnerBridgeConstructor, RunnerBridge } from "./runner-bridge";
 
 function recursiveOverrideProperty(construct: Constructor, proto: Constructor) {
     for (const key of Object.getOwnPropertyNames(proto.prototype)) {
         if (key !== 'constructor' && key !== 'destroy') {
-            construct.prototype[key] = function(this: RunnerBridge, ...args: any[]) {
+            construct.prototype[key] = function(this: RunnerBridge, ...args: JsonObject[]) {
                 return this._executeMethod(key, args);
             }
         }
